% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{test_character}
\alias{test_character}
\title{stop if x is not character}
\usage{
test_character(x)
}
\arguments{
\item{x}{variable to test}
}
\value{
an error message
}
\description{
stop if x is not character
}
