% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/useShinyLemonPie.R
\name{useShinyLemonPie}
\alias{useShinyLemonPie}
\title{Set up a Shiny app to use tarteaucitron}
\usage{
useShinyLemonPie(
  lang = "cn",
  orientation = "bottom",
  showAlertSmall = FALSE,
  cookieslist = FALSE,
  closePopup = FALSE,
  showIcon = TRUE,
  iconPosition = "BottomRight",
  iconSrc = "",
  DenyAllCta = TRUE,
  AcceptAllCta = TRUE,
  highPrivacy = TRUE,
  removeCredit = FALSE,
  moreInfoLink = TRUE,
  groupServices = FALSE,
  handleBrowserDNTRequest = FALSE,
  adblocker = FALSE,
  mandatory = TRUE,
  mandatoryCta = TRUE,
  hashtag = "#tarteaucitron",
  cookieName = "tarteaucitron",
  serviceDefaultState = "wait",
  privacyUrl = "",
  bodyPosition = "top",
  cookieDomain = "",
  readmoreLink = "",
  useExternalCss = FALSE,
  useExternalJs = FALSE,
  useBS5Css = FALSE
)
}
\arguments{
\item{lang}{Language attribute for the ShinyLemonPie input. Available values
are "ar","bg", "ca", "cn","cs", "da", "de", "el", "en", "es", "et", "fi", "fr",
"hu", "it", "ja", "lt", "lv", "nl", "no", "oc, "pl", "pt", "ro", "ru", "se",
"sk", "sv", "tr", "uk", "vi", "zh" for
Arabic, Bulgarian, Catalan, Chinese, Czech, Danish, German, Greek, English,
Spanish, Estonian, Finnish, French, Hungarian, Italian, Japanese, Lithuanian,
Latvian, Dutch, Norwegian, Occitan, Polish, Portuguese, Romanian & Moldavian,
Russian, Northern Sami, Slovak, Swedish, Turkish, Ukrainian, Vietnamese and Chinese respectively.}

\item{orientation}{Banner position (top, bottom and middle)}

\item{showAlertSmall}{logical. Show the small banner on bottom right (FALSE or TRUE)}

\item{cookieslist}{logical. Show the cookie list (FALSE or TRUE)}

\item{closePopup}{logical. Show a close X on the banner (FALSE or TRUE)}

\item{showIcon}{Show cookie icon to manage cookies}

\item{iconPosition}{Icon position (BottomRight, BottomLeft, TopRight and TopLeft)}

\item{iconSrc}{Optional: path for icon image}

\item{DenyAllCta}{logical. Show the deny all button (FALSE or TRUE)}

\item{AcceptAllCta}{logical. Show the accept all button when highPrivacy is TRUE (FALSE or TRUE)}

\item{highPrivacy}{logical. HIGHLY RECOMMANDED Disable auto consent (FALSE or TRUE)}

\item{removeCredit}{logical. Remove credit link (FALSE or TRUE)}

\item{moreInfoLink}{logical. Show more info link (FALSE or TRUE)}

\item{groupServices}{logical. Group services by category (FALSE or TRUE)}

\item{handleBrowserDNTRequest}{logical. If Do Not Track == 1, disallow all (FALSE or TRUE)}

\item{adblocker}{logical. Show a Warning if an adblocker is detected (FALSE or TRUE)}

\item{mandatory}{logical. Show a message about mandatory cookies (FALSE or TRUE)}

\item{mandatoryCta}{logical. Show the disabled accept button when mandatory on (FALSE or TRUE)}

\item{hashtag}{Open the panel with this hashtag (default : #tarteaucitron)}

\item{cookieName}{Cookie name (default : "tarteaucitron")}

\item{serviceDefaultState}{Default state (true, wait or false)}

\item{privacyUrl}{Privacy policy url}

\item{bodyPosition}{bottom or top to bring it as first element for accessibility}

\item{cookieDomain}{Shared cookie for multisite ex : .my-multisite-domaine.fr}

\item{readmoreLink}{Change the default read more link}

\item{useExternalCss}{logical. If false, the tarteaucitron.css file will be loaded}

\item{useExternalJs}{logical. If false, the tarteaucitron.js file will be loaded}

\item{useBS5Css}{logical. If true, a css theme compatible with BS5 is loaded.}
}
\value{
Scripts that ShinyLemonPie requires that are automatically inserted to the app head tag. A side effect of calling this function is that a ShinyLemonPie directory is added as a resource path using shiny::addResourcePath().
}
\description{
This function must be called from a Shiny app's UI in order for all other ShinyLemonPie functions to work.
You can call useShinyLemonPie() from anywhere inside the UI, as long as the final app UI contains the result of useShinyLemonPie()
}
\examples{
useShinyLemonPie(
  lang = "en",
  orientation = "bottom",
  showAlertSmall = FALSE,
  cookieslist = FALSE,
  closePopup = FALSE,
  showIcon = TRUE,
  iconPosition = "BottomRight",
  iconSrc = "",
  DenyAllCta = TRUE,
  AcceptAllCta = TRUE,
  highPrivacy = TRUE,
  removeCredit = FALSE,
  moreInfoLink = TRUE,
  groupServices = FALSE,
  handleBrowserDNTRequest = FALSE,
  adblocker = FALSE,
  mandatory = TRUE,
  mandatoryCta = TRUE,
  hashtag = "#tarteaucitron",
  cookieName = "tarteaucitron",
  serviceDefaultState = "wait",
  privacyUrl = "",
  bodyPosition = "top",
  cookieDomain = "",
  readmoreLink = "",
  useExternalCss = FALSE,
  useExternalJs = FALSE,
  useBS5Css = FALSE
)

}
